+++ docs/src/pages/[platform]/connected-components/storage/fileuploader/examples/Default.tsx
@@ -1,11 +1,10 @@
-import { StorageManager } from '@aws-amplify/ui-react-storage';
+import { FileUploader } from '@aws-amplify/ui-react-storage';
 
 export const App = () => {
   return (
-    <StorageManager
+    <FileUploader
       acceptedFileTypes={['image/*']}
       path="public/"
-      autoUpload={false}
       maxFileCount={1}
       isResumable
     />