+++ packages/react-auth/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aws-amplify/ui-react-auth
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [[`c3dfbe044`](https://github.com/aws-amplify/amplify-ui/commit/c3dfbe044fd18e084cd411ce0ff84dcab1a80224)]:
+  - @aws-amplify/ui@6.3.0
+  - @aws-amplify/ui-react@6.2.2
+  - @aws-amplify/ui-react-core@3.0.20
+  - @aws-amplify/ui-react-core-auth@0.0.22
+
 ## 0.0.26
 
 ### Patch Changes